+++ api/src/main/kotlin/io/bsamartins/sandbox/graphql/data/ActorRepository.kt
@@ -6,14 +6,14 @@ import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface ActorRepository : JpaRepository<Actor, Long> {
+interface ActorRepository : JpaRepository<Actor, Int> {
     fun findByName(name: String): Actor?
 }
 
 @Entity
 class Actor(
     @Id
-    val id: Long,
+    val id: Int,
 
     val name: String,
 )
