+++ api/src/main/kotlin/io/bsamartins/sandbox/graphql/data/MovieRepository.kt
@@ -14,7 +14,7 @@ import org.springframework.data.jpa.repository.JpaSpecificationExecutor
 import org.springframework.stereotype.Repository
 
 @Repository
-interface MovieRepository : JpaRepository<Movie, Long>, JpaSpecificationExecutor<Movie>, CustomMovieRepository
+interface MovieRepository : JpaRepository<Movie, Int>, JpaSpecificationExecutor<Movie>, CustomMovieRepository
 
 interface CustomMovieRepository {
     fun searchAll(query: String? = null): List<Movie>
@@ -47,7 +47,7 @@ inline fun <reified T> SearchSession.search() =
 @Indexed
 class Movie(
     @Id
-    val id: Long,
+    val id: Int,
 
     @FullTextField
     val title: String,
